refactor(account-page): move saved trips new user check to account page
