+++ lib/actions/user.js
@@ -147,8 +147,8 @@ export function createOrUpdateUser (userData, silentOnSuccess = false) {
     let requestUrl, method
 
     // Determine URL and method to use.
-    const isPost = isNewUser(loggedInUser)
-    if (isPost) {
+    const isCreatingUser = isNewUser(loggedInUser)
+    if (isCreatingUser) {
       requestUrl = `${apiBaseUrl}${API_OTPUSER_PATH}`
       method = 'POST'
     } else {
@@ -167,8 +167,8 @@ export function createOrUpdateUser (userData, silentOnSuccess = false) {
       }
 
       // Update application state with the user entry as saved
-      // (as returned) by the middleware. (Fetch trips if POST-ing.)
-      dispatch(setUser(data, isPost))
+      // (as returned) by the middleware. (Fetch trips if creating user.)
+      dispatch(setUser(data, isCreatingUser))
     } else {
       alert(`An error was encountered:\n${JSON.stringify(message)}`)
     }
