+++ lib/components/user/monitored-trip/saved-trip-list.js
@@ -10,7 +10,7 @@ import * as userActions from '../../../actions/user'
 import BackToTripPlanner from '../back-to-trip-planner'
 import AwaitingScreen from '../awaiting-screen'
 import TripSummaryPane from './trip-summary-pane'
-import { CREATE_ACCOUNT_PATH, TRIPS_PATH } from '../../../util/constants'
+import { TRIPS_PATH } from '../../../util/constants'
 import { RETURN_TO_CURRENT_ROUTE } from '../../../util/ui'
 import withLoggedInUserSupport from '../with-logged-in-user-support'
 
@@ -18,17 +18,6 @@ import withLoggedInUserSupport from '../with-logged-in-user-support'
  * This component displays the list of saved trips for the logged-in user.
  */
 class SavedTripList extends Component {
-  componentDidMount () {
-    const { loggedInUser, routeTo } = this.props
-
-    if (!loggedInUser.hasConsentedToTerms) {
-      // If a user signed up in Auth0 and did not complete the New Account wizard
-      // make the user finish set up their accounts first.
-      // monitoredTrips should not be null otherwise.
-      routeTo(CREATE_ACCOUNT_PATH)
-    }
-  }
-
   render () {
     const {trips} = this.props
     let content
@@ -74,9 +63,7 @@ const mapStateToProps = (state, ownProps) => {
   }
 }
 
-const mapDispatchToProps = {
-  routeTo: uiActions.routeTo
-}
+const mapDispatchToProps = {}
 
 export default withLoggedInUserSupport(
   withAuthenticationRequired(
