+++ lib/components/user/monitored-trip/saved-trip-screen.js
@@ -13,7 +13,7 @@ import SavedTripEditor from './saved-trip-editor'
 import TripBasicsPane from './trip-basics-pane'
 import TripNotificationsPane from './trip-notifications-pane'
 import TripSummaryPane from './trip-summary-pane'
-import { CREATE_ACCOUNT_PATH, TRIPS_PATH } from '../../../util/constants'
+import { TRIPS_PATH } from '../../../util/constants'
 import { ALL_DAYS, arrayToDayFields, WEEKDAYS } from '../../../util/monitored-trip'
 import { getActiveItineraries, getActiveSearch } from '../../../util/state'
 import { RETURN_TO_CURRENT_ROUTE } from '../../../util/ui'
@@ -102,14 +102,8 @@ class SavedTripScreen extends Component {
   }
 
   componentDidMount () {
-    const { isCreating, loggedInUser, monitoredTrips, routeTo } = this.props
-
-    if (!loggedInUser.hasConsentedToTerms) {
-      // If a user signed up in Auth0 and did not complete the New Account wizard
-      // make the user finish set up their accounts first.
-      // monitoredTrips should not be null otherwise.
-      routeTo(CREATE_ACCOUNT_PATH)
-    } else if (isCreating && hasMaxTripCount(monitoredTrips)) {
+    const { isCreating, monitoredTrips } = this.props
+    if (isCreating && hasMaxTripCount(monitoredTrips)) {
       // There is a middleware limit of 5 saved trips,
       // so if that limit is already reached, alert, then show editing mode.
       alert('You already have reached the maximum of five saved trips.\n' +
